refactor(mcp-utils): Move fake mcp server out of core into mcp-utils
